feat(EntityModel): change DatabaseKey to a public extension method
