+++ NjordinSight.EntityModel/EmptyClass.cs
@@ -4,10 +4,7 @@
 
 namespace NjordinSight.EntityModel
 {
-    /// <summary>
-    /// This class is empty to create a file to apply assembly attributes.
-    /// </summary>
-    internal class EmptyClass
-    {
-    }
+    ///// <summary>
+    ///// Empty file to apply assembly attributes.
+    ///// </summary>
 }
